+++ packages/@enotion/core/utils/index.ts
@@ -1,2 +1,3 @@
 export * from "./colors.js";
 export * from "./cn.js";
+export * from "./createContextFactory.js";