+++ packages/@enotion/hooks/src/index.ts
@@ -1,4 +1,3 @@
-export * from "./createContextFactory.js";
 export * from "./useEventListener.js";
 export * from "./useFetch.js";
 export * from "./useLocalStorage.js";